+++ dotnet/src/Extensions/TemplateEngine.Handlebars/HandlebarsPromptTemplate.cs
@@ -1,6 +1,5 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
-using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
@@ -23,7 +22,6 @@ internal class HandlebarsPromptTemplate : IPromptTemplate
         this._loggerFactory = loggerFactory ?? NullLoggerFactory.Instance;
         this._logger = this._loggerFactory.CreateLogger(typeof(HandlebarsPromptTemplate));
         this._promptModel = promptConfig;
-        this._parameters = new(() => this.InitParameters());
     }
 
     /// <inheritdoc/>
@@ -54,22 +52,6 @@ internal class HandlebarsPromptTemplate : IPromptTemplate
     private readonly ILoggerFactory _loggerFactory;
     private readonly ILogger _logger;
     private readonly PromptTemplateConfig _promptModel;
-    private readonly Lazy<IReadOnlyList<KernelParameterMetadata>> _parameters;
-
-    private List<KernelParameterMetadata> InitParameters()
-    {
-        List<KernelParameterMetadata> parameters = new(this._promptModel.InputParameters.Count);
-        foreach (var p in this._promptModel.InputParameters)
-        {
-            parameters.Add(new KernelParameterMetadata(p.Name)
-            {
-                Description = p.Description,
-                DefaultValue = p.DefaultValue
-            });
-        }
-
-        return parameters;
-    }
 
     private Dictionary<string, string> GetVariables(ContextVariables variables)
     {
@@ -89,7 +71,5 @@ internal class HandlebarsPromptTemplate : IPromptTemplate
 
         return result;
     }
-
     #endregion
-
 }
